RF: Split refactored components across different files
